Refactored build scripts
