+++ build.gradle.kts
@@ -1,63 +1,107 @@
 import org.gradle.api.tasks.testing.logging.TestExceptionFormat
-
-buildscript {
-   repositories {
-      jcenter()
-      mavenCentral()
-      maven {
-         url = uri("https://oss.sonatype.org/content/repositories/snapshots/")
-      }
-      maven {
-         url = uri("https://plugins.gradle.org/m2/")
-      }
-   }
-}
+import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
 
 plugins {
-   java
    `java-library`
-   id("java-library")
-   id("maven-publish")
    signing
-   maven
    `maven-publish`
-   kotlin("jvm").version(Libs.kotlinVersion)
+   kotlin("jvm") version "1.6.21"
 }
 
-allprojects {
-   apply(plugin = "org.jetbrains.kotlin.jvm")
+group = "io.kotest.extensions"
+version = Ci.version
 
-   group = Libs.org
-   version = Ci.version
-
-   dependencies {
-      implementation(Libs.Kotlin.stdlib)
-      implementation(Libs.Kotest.api)
-      api(Libs.Wiremock.wiremock)
-      testImplementation(Libs.Kotest.junit5)
-      testImplementation(Libs.Kotest.assertions)
+repositories {
+   mavenCentral()
+   maven {
+      url = uri("https://oss.sonatype.org/content/repositories/snapshots")
    }
+}
 
-   tasks.named<Test>("test") {
-      useJUnitPlatform()
-      testLogging {
-         showExceptions = true
-         showStandardStreams = true
-         exceptionFormat = TestExceptionFormat.FULL
-      }
+dependencies {
+   api(libs.wiremock)
+   implementation(libs.kotest.api)
+   testImplementation(libs.kotest.runner)
+   testImplementation(libs.kotest.assertions)
+}
+
+tasks.test {
+   useJUnitPlatform()
+   testLogging {
+      showExceptions = true
+      showStandardStreams = true
+      exceptionFormat = TestExceptionFormat.FULL
    }
+}
 
-   tasks.withType<org.jetbrains.kotlin.gradle.tasks.KotlinCompile> {
-      kotlinOptions.jvmTarget = "1.8"
+tasks.withType<KotlinCompile> {
+   kotlinOptions.jvmTarget = "1.8"
+}
+
+val signingKey: String? by project
+val signingPassword: String? by project
+
+val publications: PublicationContainer = (extensions.getByName("publishing") as PublishingExtension).publications
+
+signing {
+   useGpgCmd()
+   if (signingKey != null && signingPassword != null) {
+      @Suppress("UnstableApiUsage")
+      useInMemoryPgpKeys(signingKey, signingPassword)
    }
+   if (Ci.isRelease) {
+      sign(publications)
+   }
+}
 
+java {
+   withJavadocJar()
+   withSourcesJar()
+}
+
+publishing {
    repositories {
-      mavenLocal()
-      mavenCentral()
       maven {
-         url = uri("https://oss.sonatype.org/content/repositories/snapshots")
+         val releasesRepoUrl = uri("https://oss.sonatype.org/service/local/staging/deploy/maven2/")
+         val snapshotsRepoUrl = uri("https://oss.sonatype.org/content/repositories/snapshots/")
+         name = "deploy"
+         url = if (Ci.isRelease) releasesRepoUrl else snapshotsRepoUrl
+         credentials {
+            username = System.getenv("OSSRH_USERNAME") ?: ""
+            password = System.getenv("OSSRH_PASSWORD") ?: ""
+         }
+      }
+   }
+
+   publications {
+      register("mavenJava", MavenPublication::class) {
+         from(components["java"])
+         pom {
+            name.set("kotest-extensions-wiremock")
+            description.set("Kotest extension for wiremock")
+            url.set("https://www.github.com/kotest/kotest-extensions-wiremock")
+
+            scm {
+               connection.set("scm:git:http://www.github.com/kotest/kotest-extensions-wiremock")
+               developerConnection.set("scm:git:http://github.com/sksamuel")
+               url.set("https://www.github.com/kotest/kotest-extensions-wiremock")
+            }
+
+            licenses {
+               license {
+                  name.set("The Apache 2.0 License")
+                  url.set("https://opensource.org/licenses/Apache-2.0")
+               }
+            }
+
+            developers {
+               developer {
+                  id.set("sksamuel")
+                  name.set("Stephen Samuel")
+                  email.set("sam@sksamuel.com")
+               }
+            }
+         }
       }
    }
 }
-
-apply("./publish.gradle.kts")
