+++ buildSrc/build.gradle.kts
@@ -1,9 +1,7 @@
-import org.gradle.kotlin.dsl.`kotlin-dsl`
-
-repositories {
-   jcenter()
-}
-
 plugins {
    `kotlin-dsl`
+}
+
+repositories {
+   gradlePluginPortal()
 }
